Search Finish,make component clear
