+++ client/src/component/starComponent/StarComponent.js
@@ -1,23 +1,14 @@
 import axios from 'axios';
 import { useEffect, useState } from 'react';
 import base64Str from '../../constant/ImageBase64';
-import LoadMap from '../LoadMap';
-import { Container, Row, Col, Spinner, Button } from 'react-bootstrap';
-import RightClicker from '../post/RightClicker/RightClicker';
+import { Spinner } from 'react-bootstrap';
+import CustomList from '../customLIst/CustomList';
+import CustomListHiddable from '../customLIst/CustomListHiddable';
 
 export default function StarComponent() {
     const [data, setData] = useState([]);
-    const [hide, setHide] = useState(true);
-
-    const [rightClickData, setRightClickData] = useState([]);
 
     const [point, setPoint] = useState([-1, -1]);
-
-    const handleRightClick = (e, newTitle, newId) => {
-        e.preventDefault();
-        setRightClickData([newTitle, newId]);
-        setPoint([e.clientY, e.clientX]);
-    };
 
     const handlePoint = () => {
         if (point[0] !== -1 && point[1] !== -1) {
@@ -57,116 +48,13 @@ export default function StarComponent() {
         return (
             <div onClick={handlePoint}>
                 <h2>인기 로드맵</h2>
-
+                <CustomListHiddable datas={data} points={[point, setPoint]}></CustomListHiddable>
                 {/* plan을 4개씩 출력함. 그런데 idx가 3에서 더보기 버튼을 만들고, 아래는 가려진 상태로 만든다.
                 7,11이 되면 Container를 만들고  */}
-                {point[0] !== -1 && point[1] !== -1 ? <RightClicker rightClickData={rightClickData} point={point}></RightClicker> : null}
-                {data.map((_, idx) => {
-                    // 컨테이너를 만든다.
-                    if (idx === 0) {
-                        return (
-                            <Container key={idx} style={{ marginTop: '30px' }}>
-                                <Row>
-                                    {Array.from({ length: Math.min(4, data.length) }).map((_, i) => (
-                                        <Col key={data[i].plannerId}>
-                                            <div onContextMenu={(e) => handleRightClick(e, data[i].title, data[i].plannerId)}>
-                                                <LoadMap datas={data[i]}></LoadMap>
-                                            </div>
-                                        </Col>
-                                    ))}
-                                </Row>
-                            </Container>
-                        );
-                    } else if (!hide && idx % 4 === 0) {
-                        const endIdx = Math.min(idx + 3, data.length - 1);
-                        return (
-                            <Container style={{ marginTop: '30px' }}>
-                                <Row>
-                                    {Array.from({ length: endIdx - idx + 1 }).map((_, i) => (
-                                        <Col key={data[idx + i].plannerId}>
-                                            <div onContextMenu={handleRightClick}>
-                                                <LoadMap datas={data[idx + i]}></LoadMap>
-                                            </div>
-                                        </Col>
-                                    ))}
-                                    {/* 4개를 채워서 칸을 채우는것
-                                    그냥 Grid쓸껄 ㅇㅁㄻㄴㅇㄹㄴㅁㅇㄹ */}
-                                    {Array.from({ length: 4 - (endIdx - idx + 1) }).map((_, i) => (
-                                        <Col key={`empty-${i}`}></Col>
-                                    ))}
-                                </Row>
-                            </Container>
-                        );
-                    } else {
-                        return null;
-                    }
-                })}
-                {/* hide===true면 접고, false면 펼수 있도록. */}
-                {hide ? (
-                    <div>
-                        <Button
-                            className="w-25 float-end"
-                            onClick={() => {
-                                setHide(false);
-                            }}
-                        >
-                            더보기
-                        </Button>
-                    </div>
-                ) : (
-                    <div>
-                        <Button
-                            className="w-25 float-end"
-                            onClick={() => {
-                                setHide(true);
-                            }}
-                        >
-                            접기
-                        </Button>
-                    </div>
-                )}
 
                 <h2 style={{ marginTop: '50px' }}>내 로드맵</h2>
-                {data.map((_, idx) => {
-                    // 컨테이너를 만든다.
-                    if (idx === 0) {
-                        return (
-                            <Container key={idx} style={{ marginTop: '30px' }}>
-                                <Row>
-                                    {Array.from({ length: Math.min(4, data.length) }).map((_, i) => (
-                                        <Col key={data[i].plannerId}>
-                                            <div onContextMenu={(e) => handleRightClick(e, data[i].title, data[i].plannerId)}>
-                                                <LoadMap datas={data[i]}></LoadMap>
-                                            </div>
-                                        </Col>
-                                    ))}
-                                </Row>
-                            </Container>
-                        );
-                    } else if (idx % 4 === 0) {
-                        const endIdx = Math.min(idx + 3, data.length - 1);
-                        return (
-                            <Container style={{ marginTop: '30px' }}>
-                                <Row>
-                                    {Array.from({ length: endIdx - idx + 1 }).map((_, i) => (
-                                        <Col key={data[idx + i].plannerId}>
-                                            <div onContextMenu={handleRightClick}>
-                                                <LoadMap datas={data[idx + i]}></LoadMap>
-                                            </div>
-                                        </Col>
-                                    ))}
-                                    {/* 4개를 채워서 칸을 채우는것
-                                    그냥 Grid쓸껄 ㅇㅁㄻㄴㅇㄹㄴㅁㅇㄹ */}
-                                    {Array.from({ length: 4 - (endIdx - idx + 1) }).map((_, i) => (
-                                        <Col key={`empty-${i}`}></Col>
-                                    ))}
-                                </Row>
-                            </Container>
-                        );
-                    } else {
-                        return null;
-                    }
-                })}
+
+                <CustomList datas={data}></CustomList>
             </div>
         );
     }
